drivers: serial: fix compilation warning in 32-bit build
This patch fixes the following compilation warning in
32bit build.

./arch/arm/include/asm/io.h:31:43: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
   31 | #define __arch_putl(v,a)                (*(volatile unsigned int *)(a) = (v))
      |                                           ^
./arch/arm/include/asm/io.h:103:52: note: in expansion of macro '__arch_putl'
  103 | #define writel(v,c)     ({ u32 __v = v; __iowmb(); __arch_putl(__v,c); __v; })
      |                                                    ^~~~~~~~~~~
drivers/serial/serial_msm_geni.c:387:9: note: in expansion of macro 'writel'
  387 |         writel(DEF_TX_WM, base + SE_GENI_TX_WATERMARK_REG);

+++ drivers/serial/serial_msm_geni.c
@@ -311,7 +311,7 @@ static bool qcom_geni_serial_poll_bit(const struct udevice *dev, int offset,
 	return false;
 }
 
-static void qcom_geni_serial_setup_tx(u64 base, u32 xmit_size)
+static void qcom_geni_serial_setup_tx(uintptr_t base, u32 xmit_size)
 {
 	u32 m_cmd;
 
@@ -337,7 +337,7 @@ static inline void qcom_geni_serial_poll_tx_done(const struct udevice *dev)
 	writel(irq_clear, priv->base + SE_GENI_M_IRQ_CLEAR);
 }
 
-static u32 qcom_geni_serial_tx_empty(u64 base)
+static u32 qcom_geni_serial_tx_empty(uintptr_t base)
 {
 	return !readl(base + SE_GENI_TX_FIFO_STATUS);
 }
@@ -351,7 +351,7 @@ static u32 qcom_geni_serial_tx_empty(u64 base)
  * This function is used to configure the secondary sequencer with the
  * command and its associated parameters.
  */
-static inline void geni_se_setup_s_cmd(u64 base, u32 cmd, u32 params)
+static inline void geni_se_setup_s_cmd(uintptr_t base, u32 cmd, u32 params)
 {
 	u32 s_cmd;
 
@@ -362,7 +362,7 @@ static inline void geni_se_setup_s_cmd(u64 base, u32 cmd, u32 params)
 	writel(s_cmd, base + SE_GENI_S_CMD0);
 }
 
-static void qcom_geni_serial_start_tx(u64 base)
+static void qcom_geni_serial_start_tx(uintptr_t base)
 {
 	u32 irq_en;
 	u32 status;
